@@ -97,7 +97,7 @@
     </environment>
 
     <script src="~/dist/vendor.js" asp-append-version="true"></script>
-    <script src="~/dist/app.js" asp-append-version="true"></script>
+    <script src="~/dist/main.js" asp-append-version="true"></script>
 
     @RenderSection("Scripts", required: false)
 </body>
